Add palindrome test function and update test script
Added a new function, isPalindrome, to check if a word is a palindrome. This function uses another new function, reverseString, which reverses the input string. Both functions were accompanied by their respective tests. The coverage option was also included in the jest test script to ensure that all functions are well tested and to help evaluate the quality of tests.